+++ stubs/ui/command-shortcut.blade.php
@@ -1,1 +1,1 @@
-<span data-slot="command-shortcut" {{ $attributes->twMerge('text-muted-foreground ml-auto text-xs tracking-widest') }}>{{ $slot }}</span>
+<span data-slot="command-shortcut" {{ $attributes->twMerge('text-muted-foreground ms-auto text-xs tracking-widest') }}>{{ $slot }}</span>