+++ stubs/ui/dialog-header.blade.php
@@ -1,1 +1,1 @@
-<div data-slot="dialog-header" {{ $attributes->twMerge('flex flex-col gap-2 text-center sm:text-left') }}>{{ $slot }}</div>
+<div data-slot="dialog-header" {{ $attributes->twMerge('flex flex-col gap-2 text-center sm:text-start') }}>{{ $slot }}</div>